Move some definitions into separate source files to eliminate a cyclic reference that was causing some tests to fail
