+++ tsdoc/src/__tests__/ParsingBasics.test.ts
@@ -4,7 +4,8 @@ import {
   ParserContext
 } from '../index';
 import { TestHelpers } from '../parser/__tests__/TestHelpers';
-import { TSDocParserConfiguration, TSDocTagDefinition, TSDocTagSyntaxKind } from '../parser/TSDocParserConfiguration';
+import { TSDocParserConfiguration } from '../parser/TSDocParserConfiguration';
+import { TSDocTagDefinition, TSDocTagSyntaxKind } from '../parser/TSDocTagDefinition';
 
 test('01 Simple @beta and @internal extraction', () => {
   const parserContext: ParserContext = TestHelpers.parseAndMatchDocCommentSnapshot([
