+++ tsdoc/src/index.ts
@@ -1,5 +1,6 @@
 
-export { StandardTags, Standardization } from './details/StandardTags';
+export { StandardTags } from './details/StandardTags';
+export { Standardization } from './details/Standardization';
 export { StandardModifierTagSet } from './details/StandardModifierTagSet';
 export { ModifierTagSet } from './details/ModifierTagSet';
 
@@ -13,9 +14,9 @@ export { TextRange, ITextLocation } from './parser/TextRange';
 export { Token, TokenKind } from './parser/Token';
 export { TokenSequence, ITokenSequenceParameters } from './parser/TokenSequence';
 export { TSDocParser } from './parser/TSDocParser';
+export { TSDocParserConfiguration } from './parser/TSDocParserConfiguration';
 export {
   ITSDocTagDefinitionParameters,
   TSDocTagSyntaxKind,
-  TSDocTagDefinition,
-  TSDocParserConfiguration
-} from './parser/TSDocParserConfiguration';
+  TSDocTagDefinition
+} from './parser/TSDocTagDefinition';
